debug

+++ RecMST-Duplicate/src/RecMST.cu
@@ -103,12 +103,12 @@ unsigned int *d_vertex;										// starts as h_vertex
 unsigned int *d_weight;										// starts as h_weight
 unsigned int *d_edge_strength;
 unsigned int *d_edge_strength_copy;
-double* d_avg_color_r;
-double* d_avg_color_g;
-double* d_avg_color_b;
-double* d_avg_color_r_copy;
-double* d_avg_color_g_copy;
-double* d_avg_color_b_copy;
+float* d_avg_color_r;
+float* d_avg_color_g;
+float* d_avg_color_b;
+float* d_avg_color_r_copy;
+float* d_avg_color_g_copy;
+float* d_avg_color_b_copy;
 unsigned int *d_component_size;	
 unsigned int *d_old_component_size;	
 unsigned int *d_component_size_copy;	
@@ -286,12 +286,12 @@ void Init()
 	cudaMalloc( (void**) &d_component_size, sizeof(unsigned int)*no_of_vertices_orig);
 	cudaMalloc( (void**) &d_old_component_size, sizeof(unsigned int)*no_of_vertices_orig);
 	cudaMalloc( (void**) &d_component_size_copy, sizeof(unsigned int)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_r, sizeof(double)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_g, sizeof(double)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_b, sizeof(double)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_r_copy, sizeof(double)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_g_copy, sizeof(double)*no_of_vertices_orig);
-	cudaMalloc( (void**) &d_avg_color_b_copy, sizeof(double)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_r, sizeof(float)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_g, sizeof(float)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_b, sizeof(float)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_r_copy, sizeof(float)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_g_copy, sizeof(float)*no_of_vertices_orig);
+	cudaMalloc( (void**) &d_avg_color_b_copy, sizeof(float)*no_of_vertices_orig);
 
 	//Allocate memory for other arrays
 	cudaMalloc( (void**) &d_segmented_min_scan_input, sizeof(unsigned long long int)*no_of_edges_orig);
@@ -604,7 +604,7 @@ void HPGMST()
 
 	// Reweigh colors joining components so their sum when adding them up is the average
 	ReweighAndOrganizeColors<<< grid_vertexlen, threads_vertexlen, 0>>>(d_avg_color_r, d_avg_color_g, d_avg_color_b, d_component_size, d_old_component_size, d_new_supervertexIDs, no_of_vertices);
-	thrust::plus<double> binaryOp3;
+	thrust::plus<float> binaryOp3;
 	thrust::inclusive_scan_by_key(thrust::device, d_new_supervertexIDs, d_new_supervertexIDs + no_of_vertices, d_avg_color_r, d_avg_color_r_copy, binaryPred2, binaryOp3);
 	thrust::inclusive_scan_by_key(thrust::device, d_new_supervertexIDs, d_new_supervertexIDs + no_of_vertices, d_avg_color_g, d_avg_color_g_copy, binaryPred2, binaryOp3);
 	thrust::inclusive_scan_by_key(thrust::device, d_new_supervertexIDs, d_new_supervertexIDs + no_of_vertices, d_avg_color_b, d_avg_color_b_copy, binaryPred2, binaryOp3);
